Adeed main and took the menu file and added it into the main branch, then created it as a separate file
